Send progress updates using Server-Sent Events
Now instead of responding to the /download POST with a file, we send
back an html page that will stream download progress updates, then
include a download link when it's finished.

I made this change primarily out of necessity because heroku will time
out when a request takes longer than 30s to respond:
https://devcenter.heroku.com/articles/request-timeout
But it's also a nice feature to see how the download is going :)

Also now we determine the format to download by parsing the extension on
the filename. I like this better.

Also realized that my `npm run lint` command was only checking *.js
files at the root of the project, so now that's fixed.