v3 refactor: let youtube-dl create playlist dir
This simplifies the code since I don't need to read a json file to get
the name of the playlist, create the dir, and move all the files into
it. I had a lot of nice tests for doing all that but I don't need them
anymore so that's kind of bummer; job.js used to export the functions so
they could be tested but now they're just part of the Job class, which
is the only thing exported.

Also fixed lint errors