Bypass downloading to a tmp directory
This was initially done for two reasons that I can remember:
- in order to figure out what the resulting file name was (it was
  guaranteed to be the only file in the new tmp directory)
- in case the download failed then any stray files wouldn't be left
  behind (e.g. if a video file was downloaded and it failed to convert
  it to an audio file)

I don't think the second point is really an issue. And the first point
can be solved by parsing the file name from youtube-dl's output, which
is what the code now does in this commit. This means fewer fs operations
and simplifies the code a little bit too. The drawback is that if
youtube-dl changes their output messages this will break.